requirements.txt
playwright==1.49.0
playwright-stealth==1.0.6
python-telegram-bot==21.6
python-dotenv==1.0.1
pytz==2024.2
